pyproject.toml
[project]
name = "geobot-plataforma-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Álvaro Souza Oliveira",email = "66275077+alvaroxsouza@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
